Add header codes

+++ api/v1/index.php
@@ -8,16 +8,16 @@ $method = $_SERVER['REQUEST_METHOD'];
 
 $request  = explode('/', $_REQUEST['request']);
 $classname = $request[0];
-$id = $request[1];
+$id = (int)$request[1];
 $data = file_get_contents('php://input');
 $query = explode("token=", $_SERVER['QUERY_STRING']);
-
 
 try {
     spl_autoload_register(function ($class) {
         if (file_exists($class.".php")) {
             require_once $class . '.php';
         } else {
+            http_response_code(406);
             echo json_encode(
                 array('error' => "Not an available endpoint!")
                 , JSON_PRETTY_PRINT);
@@ -30,13 +30,16 @@ try {
             case "POST":
                 if (isset($query[1]) && CheckToken::isCorrectToken($query[1])) {
                     if (method_exists($request, "create") && $request->create($data)) {
+                        http_response_code(201);
                         echo $request->read($request->getId());
                     } else {
+                        http_response_code(304);
                         echo json_encode(
                             array('error' => "Could not be created!")
                             , JSON_PRETTY_PRINT);
                     }
                 } else {
+                    http_response_code(401);
                     echo json_encode(
                         array('error' => "Invalid token!")
                         , JSON_PRETTY_PRINT);
@@ -44,20 +47,24 @@ try {
                 break;
             case "PUT":
                 if (isset($query[1]) && CheckToken::isCorrectToken($query[1])) {
-                    if ($id) {
+                    if (is_int($id) && $id) {
                         if (method_exists($request, "update") && $request->update($data, $id)) {
+                            http_response_code(200);
                             echo $request->read($id);
                         } else {
+                            http_response_code(304);
                             echo json_encode(
                                 array('error' => "Could not be updated!")
                                 , JSON_PRETTY_PRINT);
                         }
                     } else {
+                        http_response_code(400);
                         echo json_encode(
                             array('error' => "No Id set!")
                             , JSON_PRETTY_PRINT);
                     }
                 } else {
+                    http_response_code(401);
                     echo json_encode(
                         array('error' => "Invalid token!")
                         , JSON_PRETTY_PRINT);
@@ -65,20 +72,24 @@ try {
                 break;
             case "DELETE":
                 if (isset($query[1]) && CheckToken::isCorrectToken($query[1])) {
-                    if ($id) {
+                    if (is_int($id) && $id) {
                         if (method_exists($request, "delete") && $request->delete($id)) {
+                            http_response_code(200);
                             echo $request->read();
                         } else {
+                            http_response_code(304);
                             echo json_encode(
                                 array('error' => "Could not be deleted!")
                                 , JSON_PRETTY_PRINT);
                         }
                     } else {
+                        http_response_code(400);
                         echo json_encode(
                             array('error' => "No Id set!")
                             , JSON_PRETTY_PRINT);
                     }
                 } else {
+                    http_response_code(401);
                     echo json_encode(
                         array('error' => "Invalid token!")
                         , JSON_PRETTY_PRINT);
@@ -86,21 +97,28 @@ try {
                 break;
             case "GET":
                 if (method_exists($request, "read")) {
-                    if ($id) {
+                    if (is_int($id) && $id) {
                         echo $request->read($id);
                     } else {
                         echo $request->read();
                     }
+                } else {
+                    http_response_code(405);
+                    echo json_encode(
+                        array('error' => "Method does not exist!")
+                        , JSON_PRETTY_PRINT);
                 }
                 break;
             default:
+                http_response_code(405);
                 echo json_encode(
-                    array('error' => "Not an available method!")
+                    array('error' => "Method not allowed!")
                     , JSON_PRETTY_PRINT);
         }
 
     }
 } catch (\Throwable $e) {
+    http_response_code(404);
     echo json_encode(
         array('exception error' => $e->getMessage())
         , JSON_PRETTY_PRINT);
